Add argument support
Refactor code into individual files.
