+++ package.json
@@ -9,6 +9,7 @@
     "typings.json"
   ],
   "dependencies": {
+    "camel-case": "^3.0.0",
     "minimist": "^1.2.0",
     "wordwrap": "^1.0.0"
   },
